+++ evaluator/functions.go
@@ -26,6 +26,8 @@ var functions = map[object.ObjectType]map[string]*object.Builtin{
 	},
 	object.INT_OBJ: {
 		"float": {Fn: intFloatFunc},
+		"abs":   {Fn: intAbsFunc},
+		"str":   {Fn: intStrFunc},
 	},
 	object.BOOL_OBJ: {},
 }